Integrate PreParserAtUserListener into PreParser directly
